:boom: Move service classes into domain-based packages
