+++ src/main/java/com/grepp/coffee/app/controller/api/admin/AdminApiController.java
@@ -1,7 +1,7 @@
 package com.grepp.coffee.app.controller.api.admin;
 
 import com.grepp.coffee.app.model.dto.CoffeeDto;
-import com.grepp.coffee.app.service.MenuService;
+import com.grepp.coffee.app.model.menu.MenuService;
 import com.grepp.coffee.infra.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
