+++ src/main/java/com/grepp/coffee/app/controller/web/order/OrderController.java
@@ -6,8 +6,8 @@ import com.grepp.coffee.app.controller.web.order.payload.OrderRequest;
 import com.grepp.coffee.app.model.dto.CoffeeDto;
 import com.grepp.coffee.app.model.dto.DetailedOrderDto;
 import com.grepp.coffee.app.model.dto.OrderDto;
-import com.grepp.coffee.app.service.MenuService;
-import com.grepp.coffee.app.service.OrderService;
+import com.grepp.coffee.app.model.menu.MenuService;
+import com.grepp.coffee.app.model.order.OrderService;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import java.util.ArrayList;
